requirements.txt
Django==1.9.5
django-filter==0.13.0
django-rest-swagger==0.3.5
djangorestframework==3.3.3
jsonfield==1.0.3
Markdown==2.6.6
psycopg2==2.6
PyYAML==3.11
wheel==0.24.0
